pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oh-my-claude-roles"
version = "0.1.2"
authors = []
description = "Claude Code 角色工具包管理器 - 从原始角色文档一键生成安装工具包"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "jsonschema>=4.0.0",
    "tenacity>=8.0.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=2.0",
    "langchain-community>=0.3.0",
]

[project.scripts]
oh-roles = "src.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=24.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --cov=src --cov-report=term --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["*/__init__.py", "tests/*"]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
docs
*/tests/*
'''

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
